Demonstration/session-management-module/local/requirements.txt
mysql-connector-python
Flask
pyjwt
cryptography